refactor: extract generation-gate confirmation flow from mcpProtocol
mcpProtocol.ts sat at 798/800 lines — the next line added anywhere would
fail check:filesize. Move the server-owned generation challenge flow
(challenge types + resolveGenerationConfirmation +
requestGenerationConfirmation + in-flight dedup map) into
mcpGateConfirmation.ts, mirroring the earlier mcpToolCatalog /
mcpSemanticGenerationFlow splits. Behavior byte-identical; the only
mechanical change is clientSupportsElicitation becoming a getter (the
flag mutates at initialize, after the factory wires up). Types re-export
from mcpProtocol so mcpSemanticGenerationFlow and tests keep their
import paths. mcpProtocol.ts: 798 → 692 lines.
